commit before checkout

+++ GamePortal/Kbalan.TouchType.Logic/Dto/StatisticDto.cs
@@ -11,7 +11,7 @@
 
         public int NumberOfGamesPlayed { get; set; }
 
-        public int AvarageSpeed { get; set; }
+        public double AvarageSpeed { get; set; }
 
     }
 }
